Add Scraped::Base class
This class contains the logic that used to be in the top level class. We
want to move it out of the top level class and into a `Base` class so
that we can make the top level `Scraped` constant into a module.

This means that the top level `Scraped` constant can now be used to add
helper methods without introducing unnecessary dependencies to
subclasses.
